backend/requirements.txt
# backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.19
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
beautifulsoup4>=4.12.2
PyPDF2>=3.0.1
python-docx>=1.0.0
aiofiles>=23.1.0